### 2022-06-18 #4488 @editorialbot
[PRE REVIEW]: py2lispIDyOM: A Python package for IDyOM
**Submitting author:** @xinyiguan (<a href="http://orcid.org/0000-0002-4570-906X">Xinyi Guan</a>)
**Repository:** https://github.com/xinyiguan/py2lispIDyOM
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
