### 2022-06-22 #4488 @faroit
@Kevin-Mattheus-Moerman @arfon I would be available to edit this but I had a brief look at the submission and I have some concerns that this is in-scope with JOSS. I am particularly worried that it doesn't pass the criteria for `substantial scholarly effort` as the main contribution is to provide a python wrapper for an existing Lisp-based software.

I do not have the knowledge in Lisp to provide a scope-check myself, specifically how much scholary effort is required to provide such a wrapper compared to a rewrite in python. Therefore I would suggest we find a Lisp expert here for a scope-check.
