### 2022-06-24 #4488 @Kevin-Mattheus-Moerman
@xinyiguan following @faroit's recommendation I've added the `query-scope` label just now which means our editorial board will inspect this submission to check if it is in scope or not. As @faroit says we'll focus on determining if it conforms to our [`substantial scholarly effort` criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). If you'd like to help us here it would be great if you could comment on what your software functionality offers on top of a wrapper of said Lisp library.
This additional scope review should take about 2 weeks to complete.
