### 2022-07-29 #4488 @Kevin-Mattheus-Moerman
@xinyiguan thanks for your patience. I have just removed the `query-scope` label since we've determined this may be in scope for JOSS. We will now proceed to try to find a handling editor for this work. Note however that unfortunately we may need to put this on a waitlist as we have a backlog of submission to process at the moment.
