### 2022-08-16 #4488 @faroit
👋  Hi @xinyiguan, thanks for your submission to JOSS, and sorry for the delay (I was on vacation last week)! I will be serving as the editor for this. We'll use this issue to identify reviewers and resolve technical issues.

@xinyiguan, I am also interested in your input on potential reviewers, so any suggestions you have are also welcome. As this toolbox is based on lisp wrapper, it would be great to have reviewers with domain knowledge for both, python and lisp.

Before you post suggestions, please check the JOSS [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and only mention them with a space between the @ and their handle (so you would refer to me as "@ faroit").
