### 2022-08-30 #4488 @xinyiguan
👋 Hi @faroit, thanks for serving as the editor for the submission. Here is a list of potential reviewers with a background in both python and lisp from the JOSS reviewer list: @ hayesall; @ aminorex; @ AoifeHughes.
