### 2022-09-05 xinyiguan/py2lispIDyOM#4 @AoifeHughes
JOSS Reviewer Comments - AH
Hello @xinyiguan,

I'll be updating this issue here with respect to review: [#4738](https://github.com/openjournals/joss-reviews/issues/4738). Most of my comments will be suggestions and so can be discussed. If anything is a major issue, I will make it clear 😊

- [x] Authors, have you made clear anywhere the contributions of the other two authors? I cannot see them listed in git commits
- [x] Would it help for install instructions to note that dependancies such as SBCL can be installed via `brew install sbcl` or, I assume, linux package mangers? - this might just help you to make an install guide fit onto your repo without directing people around (could even make an install script for users?).
- [x] Is there overlap in this project and something like: https://github.com/GuiMarion/IDyOMpy which seems to e a re-implementation? If so, perhaps it would be good to comment on this in the paper and cite?

**Please note: I will let you know when I've finished adding thoughts**
