### 2022-09-10 xinyiguan/py2lispIDyOM#4 @xinyiguan
👋 Hi @AoifeHughes, thank you so much for the feedback and comments. Here are our responses:

1. On authors' contributions: All authors provided critical feedback on the design of this project, and participated in the writing and revision of the manuscript. X.G. and C.P. conceptualized the project. X.G. and Z.R. planned the code architecture. X.G. carried out the overall computational implementation. C.P. supervised the overall project.

2. We considered adding instructions on installing SBCL and other dependent components for IDyOM such as Emacs and Quicklisp. But we decided it may be more appropriate to refer the users to the complete installation guide on the IDyOM, as these are necessary steps in the installation of the IDyOM, and the usage of this package depends on the proper installation of IDyOM on the local machine.

3. The author of the project [https://github.com/GuiMarion/IDyOMpy](https://github.com/GuiMarion/IDyOMpy) is a collaborator of the last co-author. His project of implementing IDyOM in python has not been published as a paper yet, and it is still under development, as many bugs remain to be solved. It is not entirely clear when a complete paper comparing outputs from both implementation and setting up reliable benchmarking of performance will be available. We also are not certain the extent to which alteration to the basic algorithm will attract users looking to replicate or extend research based on the original version of the algorithm. The present publication is time sensitive, as the LISP implementation is creating a bottleneck for the community that tends to only publish with main developers of the algorithms. We believe, given its popularity, that many implementations and variants of this model will bloom in the coming year. But in our view, the first step is to make the original model accessible to a larger audience of music cognition researchers.
