+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={5},
   pages={377--405},
   year={2006},
-  publisher={University of California Press USA}
+  doi = {10.1525/mp.2006.23.5.377},
+  url = {https://online.ucpress.edu/mp/article/23/5/377/62287/Expectation-in-Melody-The-Influence-of-Context-and}
 }
 
 @article{Margulis2005,
@@ -17,7 +18,8 @@
   number={4},
   pages={663--714},
   year={2005},
-  publisher={University of California Press USA}
+  doi = {10.1525/mp.2005.22.4.663},
+  url = {https://online.ucpress.edu/mp/article/22/4/663/62196/A-Model-of-Melodic-Expectation}
 }
 
 @article{Morgan2019,
@@ -27,7 +29,8 @@
   volume={189},
   pages={23--34},
   year={2019},
-  publisher={Elsevier}
+  doi = {10.1016/j.cognition.2018.12.015},
+  url = {https://www.sciencedirect.com/science/article/pii/S0010027718303317}
 }
 
 @article{Bigand2006,
@@ -38,7 +41,8 @@
   number={1},
   pages={100--130},
   year={2006},
-  publisher={Elsevier}
+  doi = {10.1016/j.cognition.2005.11.007},
+  url = {https://www.sciencedirect.com/science/article/pii/S0010027705002234}
 }
 
 @article{Eerola2009,
@@ -49,7 +53,8 @@
   number={2},
   pages={231--272},
   year={2009},
-  publisher={SAGE Publications Sage UK: London, England}
+  doi = {10.1177/102986490901300203},
+  url = {https://doi.org/10.1177/102986490901300203}
 }
 
 @article{Rohrmeier2011,
@@ -60,7 +65,8 @@
   number={2},
   pages={214--222},
   year={2011},
-  publisher={Elsevier}
+  doi = {10.1016/j.concog.2010.07.004},
+  url = {https://www.sciencedirect.com/science/article/pii/S1053810010001467}
 }
 
 @article{PearceWiggins2012,
@@ -71,7 +77,8 @@
   number={4},
   pages={625--652},
   year={2012},
-  publisher={Wiley Online Library}
+  doi = {10.1111/j.1756-8765.2012.01214.x},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1756-8765.2012.01214.x}
 }
 
 @article{Di2020,
@@ -81,7 +88,8 @@
   volume={9},
   pages={e51784},
   year={2020},
-  publisher={eLife Sciences Publications Limited}
+  doi = {10.7554/eLife.51784},
+  url = {https://elifesciences.org/articles/51784}
 }
 
 @article{Politimou2021,
@@ -91,14 +99,16 @@
   volume={203},
   pages={105020},
   year={2021},
-  publisher={Elsevier}
+  doi = {10.1016/j.jecp.2020.105020},
+  url = {https://www.sciencedirect.com/science/article/pii/S0022096520304744}
 }
 
 @phdthesis{Pearce2005,
   title={The construction and evaluation of statistical models of melodic structure in music perception and composition},
   author={Pearce, Marcus T},
   year={2005},
-  school={City University London}
+  school={City University London},
+  url = {https://openaccess.city.ac.uk/id/eprint/8459/}
 }
 
 @article{Pearce2018,
@@ -109,5 +119,6 @@
   number={1},
   pages={378--395},
   year={2018},
-  publisher={Wiley Online Library}
+  doi = {10.1111/nyas.13654},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/nyas.13654}
 }
