+++ paper/paper.md
@@ -26,62 +26,26 @@ bibliography: paper.bib
 
 # Statement of need
 
-Music is a complex, multi-layered signal that displays structures along a variety of dimensions - among which melodic
-and rhythmic sequences play a crucial role across styles and cultures [@PearceWiggins2006]. Empirical studies have
-consistently demonstrated that listeners have strong and well-defined musical predictions that reflect the long-range
-statistical regularities present in the music they have heard across their lifespan [@Margulis2005; @Morgan2019]. These
-statistics are learned through passive exposure to the music in everyday life [@Bigand2006; @Eerola2009; @Rohrmeier2011]
-.
+Music is a complex, multi-layered signal that displays structures along a variety of dimensions - among which melodic and rhythmic sequences play a crucial role across styles and cultures [@PearceWiggins2006]. Empirical studies have consistently demonstrated that listeners have strong and well-defined musical predictions that reflect the long-range statistical regularities present in the music they have heard across their lifespan [@Margulis2005; @Morgan2019]. These statistics are learned through passive exposure to the music in everyday life [@Bigand2006; @Eerola2009; @Rohrmeier2011].
 
-The Information Dynamics of Music (IDyOM) has been a well-established computational model for melodic expectation in the
-music cognition community and has been empirically tested in various behavioral and neural 
-studies [@PearceWiggins2012; @Di2020; @Politimou2021]. IDyOM models listeners’ online expectations of musical events
-using predictions from variable-order Markov chains [@Pearce2005; @Pearce2018]. Its long-term component is pre-trained on a
-large musical corpus and its short-term component dynamically learns the local statistics of a melody, simulating
-long-term learning of musical statistics and short-term learning of musical patterns respectively.
+The Information Dynamics of Music (IDyOM) has been a well-established computational model for melodic expectation in the music cognition community and has been empirically tested in various behavioral and neural 
+studies [@PearceWiggins2012; @Di2020; @Politimou2021]. IDyOM models listeners’ online expectations of musical events using predictions from variable-order Markov chains [@Pearce2005; @Pearce2018]. Its long-term component is pre-trained on a large musical corpus and its short-term component dynamically learns the local statistics of a melody, simulating long-term learning of musical statistics and short-term learning of musical patterns respectively.
 
-For each note in a melody, IDyOM outputs a probability derived by merging the long-term and short-term distributions.
-From this distribution, two information-theoretic measures characterize the predictions of the model. Surprisal (or
-Information Content) represents the expectedness (i.e. the note predicted matched the note heard) of each note given the
-long term (corpus statistics) and short term (melody statistics) context. Entropy corresponds to the degree of
-uncertainty of the prediction being made. Entropy and surprisal are intended to simulate listeners' dynamical updating
-expectations–their predictions–when listening to music.
+For each note in a melody, IDyOM outputs a probability derived by merging the long-term and short-term distributions. From this distribution, two information-theoretic measures characterize the predictions of the model. Surprisal (or Information Content) represents the expectedness (i.e. the note predicted matched the note heard) of each note given the long term (corpus statistics) and short term (melody statistics) context. Entropy corresponds to the degree of uncertainty of the prediction being made. Entropy and surprisal are intended to simulate listeners' dynamical updating expectations–their predictions–when listening to music.
 
-Although the model has been firmly established as a powerful tool to model listener’s experience, the Common Lisp
-ecosystem, in which the IDyOM model is built in, entails a significant entry barrier for researchers who intend to use
-the model. On the one hand, Common Lisp is a fairly niche programming language for data analysis and gathers a rather
[…]
