### 2022-09-18 xinyiguan/py2lispIDyOM#4 @AoifeHughes
Great, thanks for getting back on these initial points!
I should have time to finish going through the repo in the next day or so :)
