### 2022-09-27 #4738 @AoifeHughes
Hi, I would be happy to sign off on the publication of this project with one minor issue:

- [ ] Contribution guidelines

I think adding these are a requirement (ref reviewer checklist), and I think a small note on the README.md would facilitate this.

Once this point is addressed, I would support the publication.
