+++ docs/how_to_contribute.rst
@@ -1,10 +1,13 @@
 How to contribute
-=================================
+=======================
 .. toctree::
 
-Py2LispIDyOM is an ongoing project. We’d love to accept your contributions to this project.
-There are two options:
+py2lispIDyOM is an ongoing project. We’d love to accept your contributions to this project.
+You can contribute by reporting an issue or by directly contributing to the source code.
 
-1. Start with an issue report to describe any issue or suggestion.
-2. Directly contact us.
+
+1. It is generally better to start with an issue report on Github to describe any issue or suggestion you intend to fix or add.
+
+2. Follow the normal process of forking the project, setting up a new branch to work in, and making changes, as outlined in this
+`Github guideline <https://docs.github.com/en/get-started/quickstart/contributing-to-projects>`_.
 
