+++ docs/requirements.txt
@@ -1,8 +1,8 @@
-sphinx
+sphinx==5.2.2
 sphinx-rtd-theme
 myst-parser
 autoclasstoc
-sphinx-autodoc-typehints
+sphinx-autodoc-typehints==1.19.4
 pydata_sphinx_theme
 nbsphinx
 sphinx_gallery
