+++ docs/_autosummary/py2lispIDyOM.configuration.BasisOption.rst
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.BasisOption
-======================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: BasisOption
-   :members:
-
-   .. autoclasstoc::
