+++ docs/_autosummary/py2lispIDyOM.configuration.LispConvertable.rst
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.LispConvertable
-==========================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: LispConvertable
-   :members:
-
-   .. autoclasstoc::
