+++ docs/_autosummary/py2lispIDyOM.configuration.RequiredParameters.rst
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.RequiredParameters
-=============================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: RequiredParameters
-   :members:
-
-   .. autoclasstoc::
