+++ docs/_autosummary/py2lispIDyOM.configuration.StatisticalModellingParameters.rst
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.StatisticalModellingParameters
-=========================================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: StatisticalModellingParameters
-   :members:
-
-   .. autoclasstoc::
