+++ docs/_autosummary/py2lispIDyOM.configuration.RunModelConfiguration.rst
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.RunModelConfiguration
-================================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: RunModelConfiguration
-   :members:
-
-   .. autoclasstoc::
