+++ docs/_autosummary/py2lispIDyOM.configuration.DatabaseConfiguration.rst
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.DatabaseConfiguration
-================================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: DatabaseConfiguration
-   :members:
-
-   .. autoclasstoc::
