+++ docs/_autosummary/py2lispIDyOM.configuration.TrainingParameters.rst
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.TrainingParameters
-=============================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: TrainingParameters
-   :members:
-
-   .. autoclasstoc::
