+++ docs/_autosummary/py2lispIDyOM.configuration.rst
@@ -1,54 +1,0 @@
-﻿py2lispIDyOM.configuration
-==========================
-
-.. automodule:: py2lispIDyOM.configuration
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-      :toctree:
-   
-      check_recursive_typings
-      get_timestamp
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-      :toctree:
-      :template: custom-class-template.rst
-   
-      BasisOption
-      CachingParameters
-      Configuration
-      DatabaseConfiguration
-      ExperimentLogger
-      IDyOMConfiguration
-      LTMOModelOptions
-      LispConvertable
-      OutputParameters
-      Parameters
-      RequiredParameters
-      RunModelConfiguration
-      STMOModelOptions
-      StatisticalModellingParameters
-      TrainingParameters
-      ViewpointSelectionParameters
-   
-   
-
-   
-   
-   
-
-
-
