+++ docs/_autosummary/py2lispIDyOM.configuration.Configuration.rst
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.Configuration
-========================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: Configuration
-   :members:
-
-   .. autoclasstoc::
