+++ docs/_autosummary/py2lispIDyOM.configuration.CachingParameters.rst
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.CachingParameters
-============================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: CachingParameters
-   :members:
-
-   .. autoclasstoc::
