+++ _sources/_autosummary/py2lispIDyOM.configuration.ExperimentLogger.rst.txt
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.ExperimentLogger
-===========================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: ExperimentLogger
-   :members:
-
-   .. autoclasstoc::
