+++ _sources/_autosummary/py2lispIDyOM.configuration.Parameters.rst.txt
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.Parameters
-=====================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: Parameters
-   :members:
-
-   .. autoclasstoc::
