+++ _sources/_autosummary/py2lispIDyOM.configuration.check_recursive_typings.rst.txt
@@ -1,6 +1,0 @@
-py2lispIDyOM.configuration.check\_recursive\_typings
-====================================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autofunction:: check_recursive_typings
