+++ _sources/_autosummary/py2lispIDyOM.configuration.IDyOMConfiguration.rst.txt
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.IDyOMConfiguration
-=============================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: IDyOMConfiguration
-   :members:
-
-   .. autoclasstoc::
