+++ _sources/_autosummary/py2lispIDyOM.configuration.ViewpointSelectionParameters.rst.txt
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.ViewpointSelectionParameters
-=======================================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: ViewpointSelectionParameters
-   :members:
-
-   .. autoclasstoc::
