+++ _sources/_autosummary/py2lispIDyOM.configuration.OutputParameters.rst.txt
@@ -1,9 +1,0 @@
-﻿py2lispIDyOM.configuration.OutputParameters
-===========================================
-
-.. currentmodule:: py2lispIDyOM.configuration
-
-.. autoclass:: OutputParameters
-   :members:
-
-   .. autoclasstoc::
