### 2022-09-27 #4738 @xinyiguan
Hi @AoifeHughes 👋🏼

Thank you so much for your feedback. I just updated the [how to contribute](https://xinyiguan.github.io/py2lispIDyOM/how_to_contribute.html) section on the documentation page and added a small note section of "contribution guideline" on the README.md.
