### 2022-10-15 864a7f0
👷 Set `tests.yml` workflow to run on push to main
