### 2022-10-15 xinyiguan/py2lispIDyOM#8 @hayesall
Add `experiment_history` to `.gitignore`
Related to #6 and #7

The `experiment_history` directories probably should not be committed into the repository.

**Recommendation**: Enforce this through a `.gitignore`
