### 2022-10-15 xinyiguan/py2lispIDyOM#9 @hayesall
How to run the unit tests?
The `.github/workflows/tests.yml` file appears to suggest a "*Test with pytest*" step using the following:

```bash
pytest --cov=py2lispIDyOM --cov-report=xml --cov-report=html --junitxml=junit/test-results.xml
```

**Recommendation**: Clarify how tests work, and enforce this by running tests on pushes and pull requests (see #5)

## Case 1

On the [current `master` branch](https://github.com/xinyiguan/py2lispIDyOM/tree/e95cc0249a889693848d486258274b958d250f73), the tutorial notebooks work correctly. However, running this command *from the base of the repository* produces:

```python
---------- coverage: platform linux, python 3.9.13-final-0 -----------
Coverage HTML written to dir htmlcov
Coverage XML written to file coverage.xml

========================================================================================= short test summary info =========================================================================================
FAILED tests/test_export.py::TestExport::test_csv_file_check - IndexError: list index out of range
FAILED tests/test_export.py::TestExport::test_export_csv_files - IndexError: list index out of range
FAILED tests/test_export.py::TestExport::test_export_mat_files - IndexError: list index out of range
FAILED tests/test_export.py::TestExport::test_mat_file_check - IndexError: list index out of range
FAILED tests/test_extract.py::TestExtract::test_experimentinfo_extract - IndexError: list index out of range
FAILED tests/test_extract.py::TestExtract::test_melodyinfo_extract - IndexError: list index out of range
FAILED tests/test_run.py::TestIDyOMExperiment::test_case_1 - FileExistsError: [Errno 17] File exists: 'experiment_history/TestCase1/'
FAILED tests/test_run.py::TestIDyOMExperiment::test_case_2 - AssertionError: test_dataset folder is empty!
```

---

## Case 2

Running this command after switching to the `tests/` directory:

```bash
cd tests/
pytest --cov=py2lispIDyOM --cov-report=xml --cov-report=html --junitxml=junit/test-results.xml
```

results in one failed test:

```
---------- coverage: platform linux, python 3.9.13-final-0 -----------
Coverage HTML written to dir htmlcov
Coverage XML written to file coverage.xml

========================================================================================= short test summary info =========================================================================================
FAILED test_run.py::TestIDyOMExperiment::test_case_1 - FileExistsError: [Errno 17] File exists: 'experiment_history/TestCase1/'
```
