### 2022-10-15 xinyiguan/py2lispIDyOM#10 @hayesall
Code coverage estimates: clarify what it should be
Code coverage appears to be ~59% (including some cases where tests fail on my machine)

```
Name                                                                       Stmts   Miss  Cover
----------------------------------------------------------------------------------------------
/home/hayesall/Review/py2lispIDyOM/py2lispIDyOM/__init__.py            4      0   100%
/home/hayesall/Review/py2lispIDyOM/py2lispIDyOM/configuration.py     434     71    84%
/home/hayesall/Review/py2lispIDyOM/py2lispIDyOM/export.py             95     32    66%
/home/hayesall/Review/py2lispIDyOM/py2lispIDyOM/extract.py           138     45    67%
/home/hayesall/Review/py2lispIDyOM/py2lispIDyOM/run.py                62     16    74%
/home/hayesall/Review/py2lispIDyOM/py2lispIDyOM/viz.py               321    269    16%
----------------------------------------------------------------------------------------------
TOTAL                                                                       1054    433    59%
```

**Recommendation**: I don't have a hard rule for what coverage should be. I'd recommend clarifying *what coverage threshold matters to the project* in the contributing guidelines. A good portion of the library involves visualization, which tends to be slightly harder to test in my experience.
