### 2022-10-15 xinyiguan/py2lispIDyOM#12 @hayesall
Automating the installation step
The installation is pretty tricky at the moment, following from the [`idyom` installation instructions](https://github.com/mtpearce/idyom/wiki/Installation).

**Recommmendation**: Figure out how to automate installation on a common platform (e.g. Ubuntu).

## `sbcl`

This one should be pretty straightforward:

```bash
curl -L http://prdownloads.sourceforge.net/sbcl/sbcl-2.2.9-x86-64-linux-binary.tar.bz2 > sbcl-2.2.9-x86-64-linux-binary.tar.bz2
bzip2 -cd sbcl-2.2.9-x86-64-linux-binary.tar.bz2 | tar xvf -
cd sbcl-2.2.9-x86-linux
sudo sh install.sh
```

## `quicklisp`

This one might be trickier. The [`quicklisp` installation instructions](https://www.quicklisp.org/beta/) currently suggest installing in interactive mode.

I didn't test it yet (I'm less familiar with `sbcl`), but using `--eval` might be one path to follow here:

```bash
sbcl --eval '(quicklisp-quickstart:install)'
```

## `sqlite3`

This one should be easy.

```bash
sudo apt install sqlite3
```

## `IDyOM`

This one feels fairly straightforward after figuring out the username (`whoami`), and evaluating a few expressions to make sure packages are correctly downloaded.

Here is a rough draft for the fourth step:

```bash
echo '(ql:quickload "clsql")' >> .sbclrc
echo '(defun start-idyom ()' >> .sbclrc
echo "  (defvar *idyom-root* \"/home/$(whoami)/idyom/\")" >> .sbclrc
echo "  (ql:quickload \"idyom\")" >> .sbclrc
echo "  (clsql:connect '(\"/home/$(whoami)/idyom/db/database.sqlite\") :if-exists :old :database-type :sqlite3))" >> .sbclrc
```
