### 2022-10-15 #4738 @hayesall
Thank you for your patience. I've opened some issues/PRs with recommendations.

I'd definitely like to see these points addressed:

- https://github.com/xinyiguan/py2lispIDyOM/issues/12
- https://github.com/xinyiguan/py2lispIDyOM/issues/9
- https://github.com/xinyiguan/py2lispIDyOM/pull/11

If the first two items are more difficult than I expect, I will re-evaluate.

---

These points should probably be addressed, but I do not consider any of them to be blocking acceptance.

- https://github.com/xinyiguan/py2lispIDyOM/issues/7
- https://github.com/xinyiguan/py2lispIDyOM/issues/8
- https://github.com/xinyiguan/py2lispIDyOM/issues/10
- https://github.com/xinyiguan/py2lispIDyOM/pull/5
- https://github.com/xinyiguan/py2lispIDyOM/pull/6
