+++ tests/test_extract.py
@@ -10,10 +10,10 @@ from py2lispIDyOM.extract import get_song_dict_of_interest, get_all_song_dict
 
 
 class TestExtract(TestCase):
-    experiment_folder_path = 'experiment_history/25-05-22_14.10.29/'
+    experiment_folder_path = '../tests/experiment_history/25-05-22_14.10.29/'
 
     def test_experimentinfo_extract(self):
-        experiment_folder_path = 'experiment_history/25-05-22_14.10.29/'
+        experiment_folder_path = '../tests/experiment_history/25-05-22_14.10.29/'
 
         my_exp = ExperimentInfo(experiment_folder_path=experiment_folder_path)
         all_melodies = my_exp.access_melodies()
@@ -22,7 +22,7 @@ class TestExtract(TestCase):
             self.assertIsInstance(obj=all_melodies[i], cls=MelodyInfo)
 
     def test_melodyinfo_extract(self):
-        experiment_folder_path = 'experiment_history/25-05-22_14.10.29/'
+        experiment_folder_path = '../tests/experiment_history/25-05-22_14.10.29/'
         my_exp = ExperimentInfo(experiment_folder_path=experiment_folder_path)
         test_melody = my_exp.access_melodies(melody_names=['"chor-005"'])[0]
 
@@ -33,10 +33,7 @@ class TestExtract(TestCase):
 
         for idx, val in enumerate(keywords_to_match):
             self.assertIn(val, test_melody.get_idyom_output_keyword_list())
-
-            # test the func: get_idyom_output_nparray
             self.assertEqual(test_melody.get_idyom_output_nparray(val).all(), np.array(song_dict_of_interest[val]).all())
-            print(val)
 
 if __name__ == '__main__':
     unittest.main()
