+++ tests/test_export.py
@@ -13,22 +13,22 @@ import pandas as pd
 
 
 class TestExport(TestCase):
+    experiment_folder_path = './tests/experiment_history/25-05-22_14.10.29/'
 
     def test_export_mat_files(self):
-        experiment_folder_path = '../tests/experiment_history/25-05-22_14.10.29/'
+        experiment_folder_path = self.experiment_folder_path
         idyom_output_keywords = ['cpitch', 'onset', 'information.content', 'cpitch.entropy']
         Export(experiment_folder_path=experiment_folder_path,
                idyom_output_keywords=idyom_output_keywords,
                melody_names=['"chor-001"', '"chor-002"']).export2mat()
 
     def test_export_csv_files(self):
-        experiment_folder_path = '../tests/experiment_history/25-05-22_14.10.29/'
+        experiment_folder_path = self.experiment_folder_path
         Export(experiment_folder_path=experiment_folder_path,
                melody_names=['"chor-003"', '"chor-004"']).export2csv()
 
     def test_mat_file_check(self):
-        experiment_folder_path = '../tests/experiment_history/25-05-22_14.10.29/'
-
+        experiment_folder_path = self.experiment_folder_path
         chor001 = ExperimentInfo(experiment_folder_path=experiment_folder_path).melodies_dict['"chor-001"']
 
         chor001_cpitch_mat = scipy.io.loadmat(experiment_folder_path + 'outputs_in_mat/chor001_cpitch.mat')
@@ -90,7 +90,7 @@ class TestExport(TestCase):
         self.assertEqual(chor002.get_idyom_output_nparray('onset').all(), chor002_onset_mat_data.all())
 
     def test_csv_file_check(self):
-        experiment_folder_path = '../tests/experiment_history/25-05-22_14.10.29/'
+        experiment_folder_path = self.experiment_folder_path
 
         idyom_keywords_checklist = ['cpitch', 'onset',
                                     'information.content', 'entropy', 'probability']
