+++ tests/test_run.py
@@ -1,11 +1,14 @@
 import datetime
+import os
 import shutil
-import unittest
 from unittest import TestCase
 from py2lispIDyOM.run import IDyOMExperiment
 
 
+
 class TestIDyOMExperiment(TestCase):
+    bach_dataset = './tests/dataset/bach_dataset/'
+    shanx_dataset = './tests/dataset/shanx_dataset/'
 
     def test_case_1(self):
         """Test case 1:
@@ -14,8 +17,8 @@ class TestIDyOMExperiment(TestCase):
         model=stm, stmo=stmo, order_bound=5, k = 2, detail=3
         """
 
-        test_dataset_path = '../tests/dataset/bach_dataset/'
-        experiment_logger_path = '../tests/experiment_history/TestCase1/'
+        test_dataset_path = self.bach_dataset
+        experiment_logger_path = '.tests/experiment_history/TestCase1/'
         shutil.rmtree(experiment_logger_path, ignore_errors=True)
         idyom_experiment = IDyOMExperiment(test_dataset_path=test_dataset_path, experiment_logger_name='TestCase1')
         idyom_experiment.set_parameters(target_viewpoints=['cpitch'],
@@ -46,11 +49,8 @@ class TestIDyOMExperiment(TestCase):
         model=both, ltmo=ltmo, order_bound=8, k = full, detail=3
         """
 
-        test_dataset_path = '../tests/dataset/bach_dataset/'
-        pretrain_dataset_path = '../tests/dataset/shanx_dataset/'
-
-        print(test_dataset_path)
-        print(pretrain_dataset_path)
+        test_dataset_path = self.bach_dataset
+        pretrain_dataset_path = self.shanx_dataset
 
         idyom_experiment = IDyOMExperiment(test_dataset_path=test_dataset_path,
                                            pretrain_dataset_path=pretrain_dataset_path)
@@ -83,5 +83,5 @@ class TestIDyOMExperiment(TestCase):
         self.assertEqual(generated_commands, expected_commands)
 
 
-if __name__ == '__main__':
-    unittest.main()
+# if __name__ == '__main__':
+#     unittest.main()
