+++ .github/workflows/test_dev.yml
@@ -8,7 +8,29 @@ on:
   workflow_dispatch:
 
 jobs:
-  test:
+  build:
+
+#    runs-on: ${{ matrix.os }}
+#    strategy:
+#      matrix:
+#        os: [ ubuntu-latest, macos-latest]
+#        python-version: [ "3.9", "3.10" ]
+#    env:
+#      OS: ${{ matrix.os }}
+#      PYTHON: ${{ matrix.python-version }}
+#    steps:
+#      - uses: actions/checkout@v3
+#      - name: Set up Python ${{ matrix.python-version }}
+#        uses: actions/setup-python@v3
+#        with:
+#          python-version: ${{ matrix.python-version }}
+#      - name: Install dependencies
+#        run: |
+#          python -m pip install --upgrade pip
+#          pip install pytest
+#          pip install -r requirements.txt
+#      - name: Test with pytest
+#        run: pytest tests/test*
 
     runs-on: ${{ matrix.os }}
     strategy:
@@ -27,7 +49,14 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install pytest
+          pip install flake8 pytest pytest-cov coverage
           pip install -r requirements.txt
+      - name: Lint with flake8
+        run: |
+          # stop the build if there are Python syntax errors or undefined names
+          flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+          # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+          flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Test with pytest
-        run: pytest tests/test*
+        run: |
+          pytest --cov=py2lispIDyOM --cov-report=xml --cov-report=html --junitxml=junit/test-results.xml
