+++ .github/workflows/tests.yml
@@ -9,7 +9,6 @@ on:
 
 jobs:
   test:
-
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
