### 2022-10-19 281ce50
automated idyom installation process
