### 2022-10-19 0606fd0
updated README: automate IDyOM installation
