### 2022-10-22 72ead90
added new viz test
