+++ tests/test_run.py
@@ -1,13 +1,20 @@
-import datetime
-import shutil
+"""
+This test script concerns the configuration and run functionality.
+"""
+import datetime, os, shutil
 from unittest import TestCase
 from py2lispIDyOM.run import IDyOMExperiment
 
 
-
-class TestIDyOMExperiment(TestCase):
+class Test(TestCase):
     bach_dataset = './tests/dataset/bach_dataset/'
     shanx_dataset = './tests/dataset/shanx_dataset/'
+    Chabrier_krn = './tests/dataset/Chabrier_krn/'
+    Masse_krn = './tests/dataset/Masse_krn/'
+    empty_dataset = './test/dataset/empty_dataset/'
+
+    # experiment_logger_path = '.tests/experiment_history/'
+    experiment_logger_path = 'experiment_history/'  # use this path if run test locally
 
     def test_case_1(self):
         """Test case 1:
@@ -17,9 +24,13 @@ class TestIDyOMExperiment(TestCase):
         """
 
         test_dataset_path = self.bach_dataset
-        experiment_logger_path = '.tests/experiment_history/TestCase1/'
-        shutil.rmtree(experiment_logger_path, ignore_errors=True)
-        idyom_experiment = IDyOMExperiment(test_dataset_path=test_dataset_path, experiment_logger_name='TestCase1')
+        exp_folder_name = 'TestCase1'
+        experiment_logger_path = self.experiment_logger_path + exp_folder_name + '/'
+
+        if os.path.exists(experiment_logger_path):
+            shutil.rmtree(experiment_logger_path, ignore_errors=True)
+
+        idyom_experiment = IDyOMExperiment(test_dataset_path=test_dataset_path, experiment_logger_name=exp_folder_name)
         idyom_experiment.set_parameters(target_viewpoints=['cpitch'],
                                         source_viewpoints=[('cpintfref', 'cpint'), 'cpitch'],
                                         models=':stm',
@@ -27,8 +38,6 @@ class TestIDyOMExperiment(TestCase):
                                         stmo_order_bound=5,
                                         k=2,
                                         detail=3)
-
-        exp_folder_name = 'TestCase1'
 
         test_dataset_id = idyom_experiment._generate_test_dataset_id()
         generated_commands = idyom_experiment._generate_lisp_commands()
@@ -80,3 +89,175 @@ class TestIDyOMExperiment(TestCase):
                             f'(quit)'
 
         self.assertEqual(generated_commands, expected_commands)
+
+    def test_case_3(self):
+        test_dataset_path = self.Chabrier_krn
+        pretrain_dataset_path = self.Masse_krn
+
+        exp_folder_name = 'TestCase3'
+        experiment_logger_path = self.experiment_logger_path + exp_folder_name + '/'
+
+        if os.path.exists(experiment_logger_path):
+            shutil.rmtree(experiment_logger_path, ignore_errors=True)
+
+        idyom_experiment = IDyOMExperiment(test_dataset_path=test_dataset_path,
+                                           pretrain_dataset_path=pretrain_dataset_path,
+                                           experiment_logger_name=exp_folder_name)
+        idyom_experiment.set_parameters(target_viewpoints=['cpitch', 'onset'],
+                                        source_viewpoints=['cpitch', 'onset'],
+                                        models=':both',
+                                        ltmo=':ltmo',
+                                        stmo=':stmo',
+                                        ltmo_order_bound=3,
+                                        stmo_order_bound=2,
+                                        detail=2,
+                                        overwrite=True)
+
+        idyom_experiment.generate_lisp_script(write=True)
+
+        test_dataset_id = idyom_experiment._generate_test_dataset_id()
+        pretrain_dataset_id = idyom_experiment._generate_train_dataset_id()
[…]
