+++ tests/test_viz.py
@@ -1,7 +1,4 @@
-import unittest
-from unittest import TestCase, mock
-from unittest.mock import MagicMock, patch
-
+from unittest import TestCase
 import matplotlib.pyplot as plt
 
 from py2lispIDyOM.extract import MelodyInfo, ExperimentInfo
@@ -10,7 +7,6 @@ import py2lispIDyOM.viz as viz
 
 class Test(TestCase):
     experiment_folder_path = './tests/experiment_history/25-05-22_14.10.29/'
-    # experiment_folder_path = 'experiment_history/25-05-22_14.10.29/'
 
     my_exp = ExperimentInfo(experiment_folder_path=experiment_folder_path)
     all_melodies = my_exp.access_melodies()
@@ -48,18 +44,96 @@ class Test(TestCase):
         onset_in_beat_len = len(self.my_exp.melodies_dict['"chor-001"']._get_onset_beat_nparray())
         onset_answer_len = 39
 
-        chosen_ic_len = len(self.my_exp.melodies_dict['"chor-001"'].get_idyom_output_nparray(idyom_output_key='information.content'))
+        chosen_ic_len = len(
+            self.my_exp.melodies_dict['"chor-001"'].get_idyom_output_nparray(idyom_output_key='information.content'))
         ic_answer_len = 39
 
         self.assertEqual(onset_in_beat_len, onset_answer_len)
-        self.assertEqual(chosen_ic_len,ic_answer_len)
+        self.assertEqual(chosen_ic_len, ic_answer_len)
 
     def test_one_entropy_along_onsets(self):
         onset_in_beat_len = len(self.my_exp.melodies_dict['"chor-003"']._get_onset_beat_nparray())
         onset_answer_len = 34
 
-        chosen_entropy_len = len(self.my_exp.melodies_dict['"chor-003"'].get_idyom_output_nparray(idyom_output_key='entropy'))
+        chosen_entropy_len = len(
+            self.my_exp.melodies_dict['"chor-003"'].get_idyom_output_nparray(idyom_output_key='entropy'))
         entropy_answer_len = 34
 
         self.assertEqual(onset_in_beat_len, onset_answer_len)
-        self.assertEqual(chosen_entropy_len,entropy_answer_len)
+        self.assertEqual(chosen_entropy_len, entropy_answer_len)
+
+    def test_pianoroll_pitch_prediction_groundtruth(self):
+        try:
+            viz.BasicPlot.pianoroll_pitch_prediction_groundtruth(experiment_folder_path=self.experiment_folder_path,
+                                                                 starting_index=0,
+                                                                 ending_index=2,
+                                                                 showfig=False,
+                                                                 probability_colorbar=True)
+        except AssertionError:
+            self.fail('Test failed.')
+
+    def test_pianoroll_groundtruth_overall_surprisal(self):
+        try:
+            viz.BasicPlot.pianoroll_groundtruth_overall_surprisal(experiment_folder_path=self.experiment_folder_path,
+                                                                  melody_names=['"chor-005"'],
+                                                                  showfig=False)
+        except AssertionError:
+            self.fail('Test failed.')
+
+    def test_simple_plot(self):
+        try:
+            viz.BasicPlot.simple_plot(experiment_folder_path=self.experiment_folder_path,
+                                      selected_idyom_output='information.content',
+                                      melody_names=['"chor-006"'],
+                                      showfig=False)
+        except AssertionError:
+            self.fail('Test failed.')
+
+    def test_selected_surprisal_entropy(self):
+        try:
+            viz.BasicPlot.selected_surprisal_entropy(experiment_folder_path=self.experiment_folder_path,
+                                                     ic_source='information.content',
+                                                     entropy_source='entropy',
[…]
