### 2022-10-24 xinyiguan/py2lispIDyOM#12 @xinyiguan
Thank you so much for outlining the steps. I've written up a shell script to automate the IDyOM installation process (for MacOS). The instructions are also updated in the README.
