### 2022-10-24 xinyiguan/py2lispIDyOM#9 @xinyiguan
Thanks for pointing this out! There was a path issue in the tests and now it has been fixed.

Now to run the tests locally, run `python -m unittest discover` from the base of the repository, and I also enforced to run the automated tests on pushes and pull requests.
