### 2022-10-24 xinyiguan/py2lispIDyOM#8 @xinyiguan
Yes, I added the directory to `.gitignore`.
