### 2022-10-24 xinyiguan/py2lispIDyOM#7 @xinyiguan
For the modified files (plots and csv's), the contents are the same. The metadata (time created) has been changed as the previous ones are overwritten. I will add these output files to `.gitignore`.

For the data in the CSV files, the results (or the _content_ within each file) are the same. It's just the formatting being different. Internally I'm using pandas' `to_csv` for exporting data to csv file, and there is a [minor update](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.to_csv.html) after the tutorial has been written.
