### 2022-10-24 f68e0ee
Update how_to_contribute.rst
code coverage info
