+++ docs/how_to_contribute.rst
@@ -11,3 +11,6 @@ You can contribute by reporting an issue or by directly contributing to the sour
 2. Follow the normal process of forking the project, setting up a new branch to work in, and making changes, as outlined in this
 `Github guideline <https://docs.github.com/en/get-started/quickstart/contributing-to-projects>`_.
 
+
+
+Note: We expect code coverage of new features to be at least around 90%.
