### 2022-10-24 xinyiguan/py2lispIDyOM#10 @xinyiguan
Thank you for the recommendations! I added some more tests to the package, and also a note on the desired coverage threshold in the contributing guidelines. For the current coverage:

```
Name                            Stmts   Miss  Cover
---------------------------------------------------
py2lispIDyOM/__init__.py            4      0   100%
py2lispIDyOM/configuration.py     434     30    93%
py2lispIDyOM/export.py             95      6    94%
py2lispIDyOM/extract.py           138      2    99%
py2lispIDyOM/run.py                62      6    90%
py2lispIDyOM/viz.py               321     41    87%
tests/__init__.py                   0      0   100%
tests/test_export.py               75      0   100%
tests/test_extract.py              76      0   100%
tests/test_run.py                 118      0   100%
tests/test_viz.py                  63     12    81%
---------------------------------------------------
TOTAL                            1386     97    93%
```
