### 2022-10-24 xinyiguan/py2lispIDyOM#10 @hayesall
Nice! Looks good on my Ubuntu instance too.
