### 2022-10-24 xinyiguan/py2lispIDyOM#7 @hayesall
Is there a backwards-compatibility reason for keeping old versions checked into git?

Or maybe they're used in a sphinx build?

(My bad if I'm missing something, this might be a non-issue!)
