### 2022-10-24 xinyiguan/py2lispIDyOM#12 @hayesall
Nice! I do not have a mac to text on, but the setup looks good.
