### 2022-10-24 xinyiguan/py2lispIDyOM#12 @hayesall
Following up on #10: now that there's a way to automatically install dependencies, this should be incorporated into the continuous integration settings.

Right now, these lines are missed in the coverage step:

https://github.com/xinyiguan/py2lispIDyOM/blob/5d04e44be3df7a04d41cabf3a8da772a4f85a4d4/py2lispIDyOM/run.py#L113-L125

I'll be more confident about testing when everything is in CI.
