### 2022-10-26 xinyiguan/py2lispIDyOM#7 @hayesall
Okay, this makes sense! My bad, I somehow spaced that this came from a notebook in the first place.
