### 2022-10-26 e4757a2
testing idyom install CI - fix path
