+++ .github/workflows/test_dev.yml
@@ -24,8 +24,9 @@ jobs:
         uses: actions/setup-python@v3
         with:
           python-version: ${{ matrix.python-version }}
+
       - name: Run IDyOM installation shell script
-        run: ./install_idyom/install_idyom.sh
+        run: chmod +x ./install_idyom/install_idyom.sh
         shell: bash
 
       - name: Install dependencies
