+++ .github/workflows/tests.yml
@@ -12,17 +12,23 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ ubuntu-latest, macos-latest]
+        os: [ macos-latest]
         python-version: [ "3.9", "3.10" ]
     env:
       OS: ${{ matrix.os }}
       PYTHON: ${{ matrix.python-version }}
+
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v3
         with:
           python-version: ${{ matrix.python-version }}
+
+      - name: Run IDyOM installation shell script
+        run: chmod +x ./install_idyom/install_idyom.sh
+        shell: bash
+
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
