+++ .github/workflows/test_dev.yml
@@ -9,21 +9,26 @@ on:
 
 jobs:
   test:
-
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ ubuntu-latest, macos-latest]
+        os: [ macos-latest]
         python-version: [ "3.9", "3.10" ]
     env:
       OS: ${{ matrix.os }}
       PYTHON: ${{ matrix.python-version }}
+
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v3
         with:
           python-version: ${{ matrix.python-version }}
+
+      - name: Run IDyOM installation shell script
+        run: chmod +x ./install_idyom/install_idyom.sh
+        shell: bash
+
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
@@ -31,4 +36,3 @@ jobs:
           pip install -r requirements.txt
       - name: Test with pytest
         run: pytest tests/test*
-
