### 2022-10-26 xinyiguan/py2lispIDyOM#12 @xinyiguan
I just incorporated the installation to CI and added additional test to cover the `.run()` method. I changed to only the macos platform in the `tests.yml` workflow for the test because the install shell script works for (and is only tested on) macos for the moment.
