+++ paper/paper.md
@@ -44,7 +44,7 @@ To help researchers further bring insights in the music cognition domain, we int
 streamlines the research workflow for running the IDyOM model and analyzing output data. It is broadly aimed at researchers conducting IDyOM-based analysis in Python. This package makes it easier to do the following two tasks: (i) configuring and running the IDyOM model, and (ii) processing and analyzing the IDyOM output data.
 
 Users can now configure the IDyOM experiments in Python using the `IDyOMExperiment` object and, based on
-it, `py2lispIDyOM` will generate a Lisp script to run the IDyOM. To encourage an organized workflow and to improve the reproducibility of data, all data related to an experiment will be logged in a structured folder, which can be shared with other researchers to verify and replicate the experiment results. The logged data includes, but not limited to, the datasets used for the experiment, the generated Lisp script, and the IDyOM output file (in `.dat` format).
+it, `py2lispIDyOM` will generate a Lisp script to run the IDyOM. To encourage an organized workflow and to improve the reproducibility of data, all data related to an experiment will be logged in a structured folder, which can be shared with other researchers to verify and replicate the experiment results. The logged data includes, but is not limited to, the datasets used for the experiment, the generated Lisp script, and the IDyOM output file (in `.dat` format).
 
 For the processing and analysis of the IDyOM outputs, we implemented three functionalities as modules: `extract`, `export`, and `visualization`. Each module contains methods that we frequently used in our previous research projects. In the current version, we provided several common types of figures as shown below in the `visualization` module. For example, for each test melody, we can show the piano roll plots for pitch prediction distribution (\autoref{fig:Fig1}, upper panel) and visually compare it to the ground truth (\autoref{fig:Fig1}, lower panel) for this particular melody. The surprisal values can be visualized along with the corresponding piano roll, as illustrated in \autoref{fig:Fig2}. The model's output for distinct viewpoints can be visualized in \autoref{fig:Fig3} (upper panel: pitch only, middle panel: onset only, lower panel: pitch and onset combined).  \autoref{fig:Fig4} displays the predicted surprisal and entropy values (based on the pitch and onset combined model). All plotting functions in this package follow the same syntax, and therefore users can easily customize their own figures by using the provided examples. The package includes tutorials in the form of Jupyter notebooks on the GitHub repository. These tutorials present examples using the aforementioned functionalities.
 
