### 2022-11-01 #4738 @xinyiguan
Hi @faroit 👋🏼,

Thanks for the updates and the grammar check! I just merge the PR and made the new release.

The new release version is v.1.0.2.
The DOI of the Zenodo archive of the new release version is 10.5281/zenodo.7272236
