### 2022-11-08 #4738 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
